Update README.md
Bank Loan Report Project

Understanding the Main Problem

The goal of this project was to create a comprehensive Bank Loan Report to monitor and assess our bank's lending activities and performance. By providing insights into key loan-related metrics and their changes over time, the report helps in making data-driven decisions, tracking the health of the loan portfolio, and identifying trends to inform lending strategies.
Ask and Prepare
After understanding the main problem and familiarizing myself with the metadata and the meaning of each column in the dataset, I formulated specific questions to address in the subsequent phases. I used SQL to pull data from the bank’s database and established a connection between the worksheet and the database, ensuring that any changes made to the database are automatically reflected in the Excel workbook.
Data Cleaning
I cleaned the data to ensure it was free of errors, removing duplicates, filling empty cells, and resolving inconsistencies to ensure the integrity and reliability of the analysis.
Analyzing Data
With the cleaned data, I conducted Exploratory Data Analysis (EDA) to answer the questions posed in the earlier phase. Using pivot tables, I delved into the data to extract meaningful insights and trends.
Visualizing Data
I created a series of dashboards to visualize the key findings, facilitating faster information delivery and better decision-making for the management.
1.	Dashboard 1: Summary
•	Key Performance Indicators (KPIs): Total Loan Applications, Total Funded Amount, Total Amount Received, Average Interest Rate, Average Debt-to-Income Ratio (DTI).
•	Good Loan vs. Bad Loan KPIs: Categorizing loans as 'Good' (Fully Paid, Current) and 'Bad' (Charged Off) to evaluate lending performance and portfolio quality.
•	Loan Status Grid View: Providing a comprehensive overview of lending operations, including metrics such as total applications, funded amounts, received amounts, and average interest rates and DTIs by loan status.
2.	Dashboard 2: Overview
•	Monthly Trends by Issue Date (Line Chart): Showing trends in total loan applications, funded amounts, and received amounts over time.
•	Regional Analysis by State (Filled Map): Visualizing lending metrics by state to identify regional activity and disparities.
•	Loan Term Analysis (Donut Chart): Depicting loan statistics based on different loan terms.
•	Employee Length Analysis (Bar Chart): Illustrating lending metrics distributed among borrowers with varying employment lengths.
•	Loan Purpose Breakdown (Bar Chart): Providing a visual breakdown of loan metrics by loan purpose.
•	Home Ownership Analysis (Tree Map): Displaying loan metrics categorized by home ownership status for a hierarchical view.
3.	Dashboard 3: Details
•	Providing a comprehensive and user-friendly interface for accessing detailed loan data, offering a holistic snapshot of key metrics, borrower profiles, and loan performance.
By combining SQL for data extraction, Excel for data cleaning and analysis, and various visualization techniques for presenting the findings, this project effectively addresses the bank's need for a detailed and actionable Bank Loan Report.